requirements.txt
asgiref==3.2.7
Django==2.2.13
django-debug-toolbar==2.2
django-static-precompiler==1.8.2
docutils==0.16
pathtools==0.1.2
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
watchdog==0.10.2
